refactor: 모달 애니메이션 커스텀훅으로 분리
